refactor: extract Feishu bot rendering and conversation management
feishu_bot.py: 1679 → 1283 lines (-24%)

New modules:
- sjtu_agent/feishu/rendering.py (240 lines): Markdown→Feishu post conversion
  (build_post_content, parse_inline, has_table, render_table_visual, etc.)
- sjtu_agent/feishu/conversations.py (221 lines): FeishuConversationManager
  class for multi-conversation state and command dispatch

Tests updated: use FeishuConversationManager instead of direct _sessions access
