+++ sjtu_agent/feishu/__init__.py
@@ -1,0 +1,1 @@
+"""sjtu_agent.feishu — Feishu bot rendering and conversation management."""
